updated page structure
