--- pages/results/index.vue
@@ -1,10 +1,0 @@
-<template>
-  <div>
-    <h3>This is results pages</h3>
-  </div>
-</template>
-
-<script>
-export default {
-}
-</script>
